+++ R/get_ag_bulletin.R
@@ -1,8 +1,7 @@
 
-
-#' BOM agriculture bulletin information
+#' BoM agriculture bulletin information
 #'
-#'Fetch the BOM agricultural bulletin information and return a tidy data frame
+#'Fetch the BoM agricultural bulletin information and return a tidy data frame
 #'
 #' @param state Australian state or territory as postal code, see details for
 #' instruction.
@@ -21,21 +20,22 @@
 #'  }
 #'
 #' @return
-#' Data frame of a Australia BOM agricultural bulletin information.  For more
+#' Data frame of a Australia BoM agricultural bulletin information.  For more
 #' details see the vignette "Ag Bulletin Fields":
-#' \code{vignette("Ag_Bulletin Fields", package = "bomrang")} for a complete
+#' \code{vignette("Ag Bulletin Fields", package = "bomrang")} for a complete
 #' list of fields and units.
 #'
 #' @examples
+#' \dontrun{
 #' ag_bulletin <- get_ag_bulletin(state = "QLD")
-#'
+#' }
 #' @author Adam H Sparks, \email{adamhsparks@gmail.com}
 #'
 #' @references
-#' Australian Bureau of Meteorology (BOM) Weather Data Services Agriculture Bulletins
+#' Australian Bureau of Meteorology (BoM) Weather Data Services Agriculture Bulletins
 #' \url{http://www.bom.gov.au/catalogue/observations/about-agricultural.shtml}
 #'
-#' Australian Bureau of Meteorology (BOM) Weather Data Services Observation of Rainfall
+#' Australian Bureau of Meteorology (BoM) Weather Data Services Observation of Rainfall
 #' \url{http://www.bom.gov.au/climate/how/observations/rain-measure.shtml}
 #'
 #' @export
@@ -256,13 +256,12 @@ else if (state == "AUS") {
 
   tidy_df <-
     tidy_df %>%
+    dplyr::mutate_at(tidy_df, .funs = as.character, .vars = "time.zone") %>%
     dplyr::rename(
       obs_time_local = obs.time.local,
       obs_time_utc = obs.time.utc,
       time_zone = time.zone
-    ) %>%
-    dplyr::mutate_each(dplyr::funs(as.character), obs_time_utc) %>%
-    dplyr::mutate_each(dplyr::funs(as.character), time_zone)
+    )
 
   tidy_df <-
     dplyr::select(