+++ R/get_current_weather.R
@@ -1,5 +1,5 @@
 
-#' Current weather observations of a BOM station
+#' Current weather observations of a BoM station
 #'
 #' @param station_name The name of the weather station. Fuzzy string matching
 #' via \code{base::agrep} is done.
@@ -32,8 +32,9 @@
 #'   get_current_weather(latlon = c(-34, 151))
 #' }
 #' @author Hugh Parsonage, \email{hugh.parsonage@gmail.com}
-#' @import data.table
 #' @importFrom magrittr use_series
+#' @importFrom data.table :=
+#' @importFrom data.table %chin%
 #' @export get_current_weather
 
 get_current_weather <-
@@ -112,23 +113,26 @@ get_current_weather <-
         JSONurl_latlon_by_station_name %>%
         # Lat Lon are in JSON
         .[which.min(haversine_distance(lat, lon, Lat, Lon))]
-
-      on.exit(
-        message(
-          "Using station_name = '",
-          station_nrst_latlon$name,
-          "', at latitude = ",
-          station_nrst_latlon$Lat,
-          ", ",
-          "longitude = ",
-          station_nrst_latlon$Lon
+      
+      if (emit_latlon_msg) {
+        on.exit(
+          message(
+            "Using station_name = '",
+            station_nrst_latlon$name,
+            "', at latitude = ",
+            station_nrst_latlon$Lat,
+            ", ",
+            "longitude = ",
+            station_nrst_latlon$Lon
+          )
         )
-      )
+      }
 
       json_url <- station_nrst_latlon[["url"]]
     }
     if (isTRUE(httr::http_error(json_url))) {
-      stop("A station was matched but a corresponding JSON file was not found at bom.gov.au.")
+      stop("A station was matched but a corresponding JSON file was not found at
+           bom.gov.au.")
     }
 
     observations.json <-
@@ -136,7 +140,8 @@ get_current_weather <-
 
     if ("observations" %notin% names(observations.json) ||
         "data" %notin% names(observations.json$observations)) {
-      stop("A station was matched but the JSON returned by bom.gov.au was not in expected form.")
+      stop("A station was matched but the JSON returned by bom.gov.au was not in
+           expected form.")
     }
 
     # Columns which are meant to be numeric
@@ -149,8 +154,8 @@ get_current_weather <-
         "rain_trace")
     # (i.e. not raw)
     cook <- function(DT, as.DT) {
-      if (!is.data.table(DT)) {
-        setDT(DT)
+      if (!data.table::is.data.table(DT)) {
+        data.table::setDT(DT)
       }
 
       DTnoms <- names(DT)
@@ -175,7 +180,7 @@ get_current_weather <-
       }
 
       for (j in which(DTnoms %chin% double_cols)) {
-        set(DT, j = j, value = force_double(DT[[j]]))
+        data.table::set(DT, j = j, value = force_double(DT[[j]]))
       }
 
       if (!as.DT) {
@@ -191,7 +196,7 @@ get_current_weather <-
       use_series("data")
 
     if (as.data.table) {
-      setDT(out)
+      data.table::setDT(out)
     }
 
     if (raw) {